Rename the bro-related binpac files to zeek
