+++ src/bro.pac
@@ -1,4 +1,5 @@
 %extern{
+#warning "bro.pac is deprecated and will be removed in v4.1. Use zeek.pac instead."
 #include "binpac_bro.h"
 %}
