+++ src/zeek.pac
@@ -1,0 +1,10 @@
+%extern{
+#include "binpac_zeek.h"
+%}
+
+extern type ZeekAnalyzer;
+extern type ZeekVal;
+extern type ZeekPortVal;
+extern type ZeekStringVal;
+
+function network_time(): double;
